restructuring and implement docstrings   #6
